pyproject.toml
[project]
name = "figure-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langextract>=0.1.0",
    "google-generativeai>=0.3.0",
    "PyMuPDF>=1.23.0",
    "Pillow>=9.0.0",
    "python-dotenv>=1.0.0",
    "scikit-image>=0.25.2",
    "opencv-python>=4.8.0",
]
